Added new materials and textures. Updated the game logic so that bullets shoot faster and start slightly in front of the player and the camera rotates slower and sits further behind the player with an additional slight downward tilt. Changed all the lights in the game to white. Fixed a bug with bullet naming that was due to improper string/int concatination. Lastly, set shader model to 5.0 for all current pixel and vertex shaders for all configurations and platforms.

+++ DX11Starter/MenuManager.cpp
@@ -47,7 +47,7 @@ void MenuManager::DisplayGameOverMenu(DirectX::SpriteBatch * spriteBatch)
 	spriteBatch->End();
 }
 
-bool MenuManager::DetectStartClick(float xPos, float yPos)
+bool MenuManager::DetectStartClick(int xPos, int yPos)
 {
 	return ((xPos > (startButton.pos.x - (startButton.size.x / 2.f))
 		&& (xPos < startButton.pos.x + (startButton.size.x / 2.f)))
@@ -55,7 +55,7 @@ bool MenuManager::DetectStartClick(float xPos, float yPos)
 		&& (yPos < startButton.pos.y + (startButton.size.y / 2.f)))));
 }
 
-bool MenuManager::DetectQuitClick(float xPos, float yPos)
+bool MenuManager::DetectQuitClick(int xPos, int yPos)
 {
 	return ((xPos >(quitButton.pos.x - (quitButton.size.x / 2.f))
 		&& (xPos < quitButton.pos.x + (quitButton.size.x / 2.f)))
